[Model Refactoring] Rename deepseek_v4.py to model.py
