--- vllm/models/deepseek_v4/amd/deepseek_v4.py
@@ -1,1 +1,0 @@
-../nvidia/deepseek_v4.py
