+++ vllm/models/deepseek_v4/amd/model.py
@@ -1,0 +1,1 @@
+../nvidia/model.py
