+++ vllm/models/deepseek_v4/amd/mtp.py
@@ -1,0 +1,1 @@
+../nvidia/mtp.py
